add error catching

+++ entrypoint.sh
@@ -73,7 +73,16 @@ echo "${INPUT_KUBECONFIG}" > ~/.kube/config
 fi
 
 echo -e "\033[36mExecuting tkn\033[0m"
-tkn task start --showlog ${PTARG} ${SAARG} -n ${INPUT_NAMESPACE} ${INPUT_TASK} $INPUT_ARGS
+
+if ! tkn task start --showlog ${PTARG} ${SAARG} -n ${INPUT_NAMESPACE} ${INPUT_TASK} $INPUT_ARGS; then
+    echo "some_command returned an error"
+    exit 1
+fi
+
+# if <condition> ; then
+#   echo "Game over!"
+#   exit 1
+# fi
 
 echo -e "\033[36mCleaning up: \033[0m"
 rm ./run.sh -Rf
